Add files via upload
This is a simple and interactuve Quiz Game project developed in Java. It demonstrates the basic concepts of java like switch case,Methods,Conditional statements etc. It keeps track of the user's score and displays the result at the end of the quiz. It also allows users to restart the quiz after completion.